pyproject.toml
[project]
name = "python-wa-discord-bot"
version = "0.1.0"
description = "Discord bot for WeakAuras community support"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Leon Solis", email = "krazyito65@gmail.com" },
]
keywords = ["discord", "bot", "weakauras", "world-of-warcraft"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
    "Topic :: Games/Entertainment",
]
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.3",
    "pyyaml>=6.0.2",
    "django>=5.1.0",
    "django-allauth>=65.3.0",
    "requests>=2.32.0",
    "djangorestframework>=3.15.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.12.12",
    "sphinx>=8.2.3",
    "sphinx-autodoc-typehints>=3.2.0",
    "sphinx-rtd-theme>=3.0.2",
    "uv>=0.8.15",
]
